[Messenger] make StopWorkerOnSignalsListener listen by default on SIGTERM and SIGINT
